+++ pkg/api/cloud/ssh/user/delete.go
@@ -1,0 +1,34 @@
+package user
+
+import (
+	"context"
+	"net/url"
+
+	"github.com/sitehostnz/gosh/pkg/utils"
+)
+
+// Delete an existing SSH user.
+func (s *Client) Delete(ctx context.Context, request DeleteRequest) (response DeleteResponse, err error) {
+	uri := "cloud/ssh/user/delete.json"
+	keys := []string{
+		"client_id",
+		"server_name",
+		"username",
+	}
+
+	values := url.Values{}
+	values.Add("client_id", s.client.ClientID)
+	values.Add("server_name", request.ServerName)
+	values.Add("username", request.Username)
+
+	req, err := s.client.NewRequest("POST", uri, utils.Encode(values, keys))
+	if err != nil {
+		return response, err
+	}
+
+	if err := s.client.Do(ctx, req, &response); err != nil {
+		return response, err
+	}
+
+	return response, nil
+}