+++ pkg/api/ssh/key/create.go
@@ -2,6 +2,7 @@ package key
 
 import (
 	"context"
+	"fmt"
 	"net/url"
 
 	"github.com/sitehostnz/gosh/pkg/utils"
@@ -20,7 +21,7 @@ func (s *Client) Create(ctx context.Context, opts CreateRequest) (response Creat
 	values := url.Values{}
 	values.Add("label", opts.Label)
 	values.Add("content", opts.Content)
-	values.Add("params[custom_image_access]", opts.CustomImageAccess)
+	values.Add("params[custom_image_access]", fmt.Sprint(utils.BoolToInt(opts.CustomImageAccess)))
 
 	req, err := s.client.NewRequest("POST", u, utils.Encode(values, keys))
 	if err != nil {