+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v0.3.x] - 2024-05-14
+### Added
+- list/get/update/remove for `/cloud/db`
+- list/get/update/remove for `/cloud/db/user`
+- list/get/update/remove for `/cloud/db/grants`
+- cloud stack image list_all `/cloud/image`
+- add user/add/edit/update and ssh key 
+
+### Fixes
+- version updates
+- correct the ssh key update params
+- add wrappers around api return types to get bools, not ints.
+
 ## [v0.3.4] - 2024-03-12
 ### Added
 - Add ability to update SSH Keys.